Extra Stuff with Health, EnemySpanwer,and Controller
Added a publick EnemySpawner object to get rid of an error.

Need to fix EnemySpawner

Made CurrentHealth public and a NetworkVariable<Float>. Want to make this readonly, and then add an "add to health" funciton in health, and have potion call it.

Potion: added .value to CurrentHealth. Just need to refactor that when we make the new function for "adidng health"